Testing GitHub
Testing code written in 2011 for a game.
Code is not important, just learning how to use GitHub. 

Purpose of the script is just to "zoom-in" so players have a"bigger" player model for you to aim at.
Added some extra fun mini-scripts such as a jump crouch. Very useful for Source/VALVe made games.